refactor: move enable location screen to addresses directory
