+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/addresses/enableLocationScreen/EnableLocationScreenUIEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.enableLocationScreen
+package net.thechance.mena.identity.presentation.screen.addresses.enableLocationScreen
 
 sealed interface EnableLocationScreenUIEffect {
     data object NavigateBack : EnableLocationScreenUIEffect
