+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/addresses/pickLocation/PickLocationScreen.kt
@@ -22,7 +22,7 @@ import net.thechance.mena.identity.presentation.base.BaseScreen
 import net.thechance.mena.identity.presentation.components.AuthAppBar
 import net.thechance.mena.identity.presentation.components.ErrorSnackBar
 import net.thechance.mena.identity.presentation.screen.addresses.myAddresses.AddressUIState
-import net.thechance.mena.identity.presentation.screen.enableLocationScreen.EnableLocationScreen
+import net.thechance.mena.identity.presentation.screen.addresses.enableLocationScreen.EnableLocationScreen
 import net.thechance.mena.identity.presentation.screen.addresses.pickLocation.components.EditMapButton
 import net.thechance.mena.identity.presentation.screen.addresses.pickLocation.components.GpsFabButton
 import net.thechance.mena.identity.presentation.screen.addresses.pickLocation.components.PickLocationMap
